Improved visitor behavior, also modified payVisitFee method in EconomySystem
