+++ assets/prefabs/Gooeys/redGooey.prefab
@@ -1,7 +1,7 @@
 {
   "alwaysRelevant": true,
   "DisplayName": {
-    "name": "Red Gooey"
+    "name": "redGooey"
   },
   "persisted": true,
   "Color": {
