--- assets/prefabs/Gooeys/yellowgooey.prefab
@@ -1,135 +1,0 @@
-{
-  "alwaysRelevant": true,
-  "DisplayName": {
-    "name": "Yellow Gooey"
-  },
-  "persisted": true,
-  "Color": {
-    "color": [
-      191,
-      65,
-      65,
-      255
-    ]
-  },
-  "skeletalmesh": {
-    "mesh": "gooey",
-    "heightOffset": -2.0,
-    "material": "yellowGooeySkin",
-    "animationPool": [
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyLookRightLeft",
-      "gooeyLookRightLeft",
-      "gooeyLookRightUp"
-    ],
-    "loop": true,
-    "scale": [
-      0.25,
-      0.25,
-      0.25
-    ]
-  },
-  "Behavior": {
-    "tree": "GooKeeper:gooey"
-  },
-  "Stand": {
-    "animationPool": [
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyIdleBreathing",
-      "gooeyLookRightLeft",
-      "gooeyLookRightLeft",
-      "gooeyLookRightUp"
-    ]
-  },
-  "Walk": {
-    "animationPool": [
-      "gooeyIdleBreathing"
-    ]
-  },
-  "location": {},
-  "Character": {},
-  "AliveCharacter": {},
-  "CharacterMovement": {
-    "groundFriction": 16,
-    "speedMultiplier": 0.3,
-    "distanceBetweenFootsteps": 0.2,
-    "distanceBetweenSwimStrokes": 2.5,
-    "height": 1.0,
-    "radius": 0.5,
-    "jumpSpeed": 12
-  },
-  "Network": {},
-  "MinionMove": {},
-  "Health": {},
-  "BoxShape": {
-    "extents": [
-      1.0,
-      1.0,
-      1.0
-    ]
-  },
-  "Trigger": {
-    "detectGroups": [
-      "engine:debris",
-      "engine:sensor"
-    ]
-  },
-  "Flee": {},
-  "FleeOnHit": {},
-  "FindNearbyPlayers": {
-    "searchRadius": 15
-  },
-  "Follow": {
-    "entityToFollow": EntityRef.NULL
-  },
-  "NPCMovement": {},
-  "Gooey": {
-    "profitPayOff": 3,
-    "biome": ["FOREST", "PLAINS"],
-    "blockBelow": "core:Grass",
-    "SPAWN_CHANCE": 35,
-    "MAX_GROUP_SIZE": 4
-  },
-  "Friendly": {
-     "friendlinessFactor": 85
-  }
-}
