+++ assets/prefabs/Pen Blocks/bluepen.prefab
@@ -1,6 +1,6 @@
 {
     "PenBlock": {
-        "type": "Blue Gooey",
+        "type": "blueGooey",
         "cutoffFactor": 6
     }
 }
