+++ assets/prefabs/Pen Blocks/redpen.prefab
@@ -1,6 +1,6 @@
 {
     "PenBlock": {
-        "type": "Red Gooey",
+        "type": "redGooey",
         "cutoffFactor": 1
     }
 }
