+++ assets/prefabs/Pen Blocks/yellowpen.prefab
@@ -1,6 +1,6 @@
 {
     "PenBlock": {
-        "type": "Yellow Gooey",
+        "type": "yellowGooey",
         "cutoffFactor": 3
     }
 }
