+++ deltas/prefabs/redGooey.prefab
@@ -1,0 +1,5 @@
+{
+  "Behavior": {
+    "tree": "GooKeeper:capturedGooey"
+  }
+}
