+++ src/main/java/org/terasology/gookeeper/component/VisitBlockComponent.java
@@ -22,7 +22,23 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class VisitBlockComponent implements Component {
+    /**
+     * Boolean whether the block is currently occupied by a visitor
+     */
     public boolean isOccupied = false;
+
+    /**
+     * The type of pen to which the visit block is attached
+     */
     public String type = "undefined";
+
+    /**
+     * The cutoff factor for the visit block depending upon the rarity of the housed gooeys
+     */
     public float cutoffFactor;
+
+    /**
+     * The number of gooeys housed in the corresponding pen
+     */
+    public int gooeyQuantity = 0;
 }
